file: new file for new file builder feature
